added sonarcloud exclusions refactored the code
