+++ backend/src/main/java/org/eclipse/tractusx/semantics/hub/AspectModelService.java
@@ -24,10 +24,7 @@ import org.springframework.http.ResponseEntity;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.google.common.collect.Lists;
 
-import org.eclipse.esmf.aspectmodel.resolver.services.VersionedModel;
 import org.eclipse.esmf.aspectmodel.urn.AspectModelUrn;
-import org.eclipse.esmf.metamodel.Aspect;
-import org.eclipse.tractusx.semantics.hub.bamm.SDKAccessHelper;
 import org.eclipse.tractusx.semantics.hub.persistence.PersistenceLayer;
 import io.vavr.control.Try;
 import org.eclipse.tractusx.semantics.hub.api.ModelsApiDelegate;
